using variables in report
