+++ generate_report/views.py
@@ -4,7 +4,7 @@ from reportlab.lib.pagesizes import A4
 from django.http import HttpResponse
 from pygithub_api_integration.models import Repository
 from pygithub_api_integration.models import Contributor
-from generate_report.models import Report
+from pygithub_api_integration.models import Report
 
 
 def pdfView(request, repo_id):
@@ -15,12 +15,12 @@ def pdfView(request, repo_id):
     # commiters = Contributor.objects.filter(repository=repo_id)
 
     response = HttpResponse(content_type='application/pdf')
-    report_name = 'attachment; filename=Relatorio_' + repo.name + '.pdf'
+    report_name = 'attachment; filename=Relatorio_' + report.repo_name + '.pdf'
     response['Content-Disposition'] = report_name
 
     pdf = canvas.Canvas(response, pagesize=A4)
 
-    pdf.drawString(100, 100, "Hello world.")
+    pdf.drawString(20, 50, report.repo_name)
 
     pdf.showPage()
     pdf.save()
